Phase 3: supervised vs pretrained SAE comparison (v6.0)
Fix 1: pretrained GemmaScope SAE loader R²=-6.85 → route evaluate.py
through inventory.load_sae() wrapper which follows the documented JumpReLU
convention (no b_dec subtraction in encoder).

Fix 2: bf16 log_softmax underflow producing NaN KL for comma/period
(the two highest-frequency features). Added .float() cast before
log_softmax on causal.py:471 and :506.

Three new experiments, each a 3-way comparison across (S) supervised,
(U) unsupervised-within-our-SAE, and (P) pretrained GemmaScope latents.
The (S) vs (U) gap isolates supervision as the causal factor (same
architecture, same training data, same loss minus supervision term).

- feature_splitting.py (--step splitting): coverage / specificity /
  greedy N@80% per-feature for 10 targets
- circuit.py (--step circuit): closing-bracket prediction with
  attribution patching (fp32 matmul, multi-token-variant target)
- intervention.py (--step intervention): single-latent ablation KL
  at positive vs negative positions, targeting_ratio metric

Robustness: target-feature selection falls back to top-k-by-positives if
hardcoded hints don't match the catalog; max_n in set cover scales with
pool size; explicit dtype casts throughout; degenerate ratio handling;
bounds checks on feat_idx; empty-result guards.

+++ pipeline/causal.py
@@ -468,7 +468,10 @@ def test_feature_necessity(model, sae, cfg):
             logits = model.run_with_hooks(
                 toks, fwd_hooks=[(cfg.hook_point, full_recon_hook)]
             )
-            recon_logprobs.append(F.log_softmax(logits[0].cpu(), dim=-1))
+            # Cast to fp32 before log_softmax to avoid bf16 underflow on
+            # high-frequency features (comma, period) — without this, the
+            # log_softmax→KL computation produces NaN for Gemma (bf16 model).
+            recon_logprobs.append(F.log_softmax(logits[0].float().cpu(), dim=-1))
 
     # Per-feature ablation
     results = []
@@ -503,7 +506,7 @@ def test_feature_necessity(model, sae, cfg):
 
                 active_pos = active_mask[i].nonzero(as_tuple=True)[0]
                 base_lp = recon_logprobs[i][active_pos]
-                abl_lp = F.log_softmax(abl_logits[0, active_pos].cpu(), dim=-1)
+                abl_lp = F.log_softmax(abl_logits[0, active_pos].float().cpu(), dim=-1)
 
                 kl = (base_lp.exp() * (base_lp - abl_lp)).sum(dim=-1)
                 kl_sum += kl.clamp(min=0).sum().item()
